requirements.txt
streamlit>=1.28.0
google-genai>=1.20.0
streamlit-quill>=0.9.0
pandas>=1.5.0